Split entities tests into another file
